requirements.txt
bagit==1.7.0
CacheControl==0.11.7
certifi==2018.11.29
chardet==3.0.4
cwltool==1.0.20181201184214
decorator==4.3.0
html5lib==1.0.1
idna==2.7
isodate==0.6.0
lockfile==0.12.2
lxml==4.2.5
mistune==0.8.4
mypy-extensions==0.4.1
networkx==2.2
prov==1.5.1
psutil==5.4.8
pyparsing==2.3.0
python-dateutil==2.7.5
rdflib==4.2.2
rdflib-jsonld==0.4.0
requests==2.20.1
ruamel.yaml==0.15.77
scandir==1.9.0
schema-salad==3.0.20181129082112
shellescape==3.4.1
six==1.11.0
subprocess32==3.5.3
typing==3.6.6
typing-extensions==3.6.6
urllib3==1.24.1
webencodings==0.5.1
